+++ python/metatensor-learn/CHANGELOG.md
@@ -17,6 +17,8 @@ follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 -->
 
+## [Version 0.2.2](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-learn-v0.2.2) - 2024-05-16
+
 ### Added
 
 - Added torch-style activation function module maps to `metatensor.learn.nn`: `ReLU`,
@@ -26,6 +28,7 @@ follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   and `InvariantTanh` (#513)
 
 ### Fixed
+
 - Set correct device for output of when torch default device is different than input device (#595)
 
 ## [Version 0.2.1](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-learn-v0.2.1) - 2024-03-01